part 2 of rewrite
