+++ commands/image/balls.js
@@ -1,10 +1,9 @@
 const Discord = module.require('discord.js');
 
-
-module.exports.run = async (client, message, args) => {
+module.exports = {
+  commands: 'balls' ,
+  callback: (message, arguments, text) => {
     const attachment =  new Discord.MessageAttachment('./content/politics.gif', 'politics.gif')
     message.channel.send(attachment)
-}
-module.exports.help = {
-  name: "balls"
+  },
 }
